Cleaned up code, more efficiency, build saving system
Cleaned up a lot of the code around building,
made it more efficient and modular for future additions

Created an auto save/load for obj/Can_Build items in the world.

Instead of adding crafting items in New(), it does it when the world
loads, but before it loads all the custom crafting tables as not to dupe
inventory.