Update for a couple things
Added a max items define set to 12.
To do sacks, just put the objects into the sacks directly, no messing
with a player defined maxitems

Enemies don't walk diag.

Made it so whoever dealt most damage to an enemy gets the exp and the
item for 60 seconds.

Fixed bag opens (surprised it even worked) and how a bag closes when you
craft something